# specs/001-naaseh-v1-baseline/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Na'aseh v1 API
  version: 1.0.0
  description: Same-origin API for authentication, administration, tasks, groups, sync, and reminders.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /categories/{categoryId}:
    parameters:
      - name: categoryId
        in: path
        required: true
        schema: { $ref: '#/components/schemas/Ulid' }
    patch:
      tags: [Categories, Administration]
      operationId: updateCategory
      description: Requires an active application user with role `admin`.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IfMatchVersion'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/CategoryPatch' }
      responses:
        '200':
          description: Category updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Categories, Administration]
      operationId: archiveCategory
      description: Requires an active application user with role `admin`; historical task references remain valid.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IfMatchVersion'
      responses:
        '200':
          description: Category archived.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-naaseh
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 32, maxLength: 256 }
...
    IfMatchVersion:
      name: If-Match
      in: header
      required: true
      schema: { type: string, pattern: '^"[0-9]+"$' }
...
  responses:
...
    Forbidden:
      description: Authenticated but not authorized.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found or deliberately concealed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: Optimistic concurrency or uniqueness conflict.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ConflictProblem' }
...
  schemas:
    Ulid:
      type: string
      pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
...
    CategoryCreate:
      type: object
      additionalProperties: false
      required: [name, color]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        defaultAssigneeId: { oneOf: [{ $ref: '#/components/schemas/Ulid' }, { type: 'null' }] }
        color: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
    CategoryPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        defaultAssigneeId: { oneOf: [{ $ref: '#/components/schemas/Ulid' }, { type: 'null' }] }
        color: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
    Category:
      allOf:
        - $ref: '#/components/schemas/CategoryCreate'
        - type: object
          required: [categoryId, status, version]
          properties:
            categoryId: { $ref: '#/components/schemas/Ulid' }
            status: { type: string, enum: [active, archived] }
            version: { type: integer, minimum: 1 }
...
    Problem:
      type: object
      required: [type, title, status, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        correlationId: { type: string }
        fieldErrors: { type: object }
    ConflictProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            currentVersion: { type: integer }
            current: { type: object }
            conflictingFields: { type: array, items: { type: string } }